[frontend] refactor: rename LicenceBanner component
